quadcopter-simulation/requirements.txt
PyOpenGL
numpy
scipy
